move files around;TODO fix datepicker is not a function
