+++ app/components/App/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import SidePanel from '../SidePanel/SidePanel';
 import Header from '../Header/Header';
-import ScheduleWizard from '../ScheduleWizard/ScheduleWizard';
+import ScheduleWizard from '../Common/ScheduleWizard';
 
 class App extends Component {
   render() {
