+++ app/components/ScheduleWizard/index.js
@@ -1,1 +1,5 @@
-export {default} from './ScheduleWizard'
+export {ScheduleWizard} from './ScheduleWizard'
+export {TimeSettings} from './TimeSettings'
+export {InfoSettings} from './InfoSettings'
+export {BountySettings} from './BountySettings'
+export {ConfirmSettings} from './ConfirmSettings'
